Add script to create source releases (#9)
## Motivation for the change, related issues

We are doing some private releases to be used by the Studio app but want
to release the source in respect of the GPL license.

## Implementation details

This PR:
- Updates `.gitattributes` to avoid exporting large, non-source files
via `git archive`. This change reduced the size of the source package
from 810MB to 81MB (oddly by an exact factor of 10).
- Adds a script for running `git archive` and manually adding a couple
of supporting items like the `isomorphic-git` submodule and build
instructions for the source package.

## Testing Instructions (or ideally a Blueprint)

From the project root, run:
`./tools/scripts/package-source.sh <branch-name> <output-tarball-path>`

Find a place to unpack the tarball, examine it, open
`BUILD-FROM-SOURCE-RELEASE.md`, and follow the instructions to rebuild
php-wasm and wordpress-builds. This will likely take a long time. But
once you've done the setup, it should be possible to build specific
packages we release to NPM using `npx nx` commands.