Update private Playground instance to use dedicated CORS proxy (#56)
## Motivation for the change, related issues

The public CORS proxy does not support CORS for the private Playground
origin. We have configured a private CORS proxy to support that origin,
and we need to configure the private Playground instance to use that
proxy.

## Implementation details

This PR updates the private Playground deployment workflow to specify a
CORS proxy URL configured in the GitHub Actions environment.

## Testing Instructions (or ideally a Blueprint)

To save time , I am just going to deploy this update because this is a
simple change that I believe should work. I will test a deployment
immediately after merging to confirm.